fastapi_service/pyproject.toml
[project]
name = "daily-tech-podcast"
version = "0.1.0"
description = "Create podcast by n8n"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi[standard]>=0.116.1",
    "feedparser>=6.0.11",
    "mutagen>=1.47.0",
    "podgen>=1.1.0",
    "pygithub>=2.6.1",
]

[tool.uv.workspace]
members = [
    "automation",
]
